+++ firedrake/adjoint_utils/blocks/function.py
@@ -5,7 +5,7 @@ from ufl.formatting.ufl2unicode import ufl2unicode
 from ufl.algorithms.analysis import extract_arguments_and_coefficients
 from pyadjoint import Block, OverloadedType, AdjFloat
 import firedrake
-from firedrake.adjoint.checkpointing import maybe_disk_checkpoint, \
+from firedrake.adjoint_utils.checkpointing import maybe_disk_checkpoint, \
     DelegatedFunctionCheckpoint
 from .backend import Backend
