+++ firedrake/adjoint_utils/projection.py
@@ -1,6 +1,6 @@
 from functools import wraps
 from pyadjoint.tape import annotate_tape, stop_annotating, get_working_tape
-from firedrake.adjoint.blocks import ProjectBlock, SupermeshProjectBlock
+from firedrake.adjoint_utils.blocks import ProjectBlock, SupermeshProjectBlock
 from firedrake import function
 from ufl.domain import extract_unique_domain
