+++ firedrake/solving.py
@@ -25,7 +25,7 @@ import firedrake.linear_solver as ls
 import firedrake.variational_solver as vs
 from firedrake import dmhooks, function, solving_utils, vector
 import firedrake
-from firedrake.adjoint import annotate_solve
+from firedrake.adjoint_utils import annotate_solve
 from firedrake.petsc import PETSc
 from firedrake.utils import ScalarType
