+++ tests/regression/test_adjoint_operators.py
@@ -14,12 +14,11 @@ def handle_taping():
 
 @pytest.fixture(autouse=True, scope="module")
 def handle_annotation():
-    from firedrake_adjoint import annotate_tape, continue_annotation
+    from firedrake.adjoint import annotate_tape, continue_annotation
     if not annotate_tape():
         continue_annotation()
     yield
-    # Since importing firedrake_adjoint modifies a global variable, we need to
-    # pause annotations at the end of the module
+    # Ensure annotation is paused when we finish.
     annotate = annotate_tape()
     if annotate:
         pause_annotation()
@@ -47,7 +46,7 @@ def vector(request):
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_constant():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(10, 10)
     V1 = FunctionSpace(mesh, "CG", 1)
 
@@ -63,7 +62,7 @@ def test_interpolate_constant():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_with_arguments():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(10, 10)
     V1 = FunctionSpace(mesh, "CG", 1)
     V2 = FunctionSpace(mesh, "CG", 2)
@@ -84,7 +83,7 @@ def test_interpolate_with_arguments():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_scalar_valued():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = IntervalMesh(10, 0, 1)
     V1 = FunctionSpace(mesh, "CG", 1)
     V2 = FunctionSpace(mesh, "CG", 2)
@@ -111,7 +110,7 @@ def test_interpolate_scalar_valued():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_vector_valued():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(10, 10)
     V1 = VectorFunctionSpace(mesh, "CG", 1)
     V2 = VectorFunctionSpace(mesh, "DG", 0)
@@ -133,7 +132,7 @@ def test_interpolate_vector_valued():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_tlm():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(10, 10)
     V1 = VectorFunctionSpace(mesh, "CG", 1)
     V2 = VectorFunctionSpace(mesh, "DG", 0)
@@ -161,7 +160,7 @@ def test_interpolate_tlm():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_tlm_wit_constant():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = IntervalMesh(10, 0, 1)
     V1 = FunctionSpace(mesh, "CG", 2)
     V2 = FunctionSpace(mesh, "DG", 1)
@@ -195,7 +194,7 @@ def test_interpolate_tlm_wit_constant():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_bump_function():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(10, 10)
     V = FunctionSpace(mesh, "CG", 2)
 
@@ -213,7 +212,7 @@ def test_interpolate_bump_function():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_self_interpolate():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(1, 1)
     V = FunctionSpace(mesh, "CG", 1)
     u = Function(V)
@@ -230,7 +229,7 @@ def test_self_interpolate():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_self_interpolate_function():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(1, 1)
     V = FunctionSpace(mesh, "CG", 1)
     u = Function(V)
@@ -248,7 +247,7 @@ def test_self_interpolate_function():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_interpolate_to_function_space():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(1, 1)
     V = FunctionSpace(mesh, "CG", 1)
     W = FunctionSpace(mesh, "DG", 1)
@@ -271,7 +270,7 @@ def test_interpolate_hessian_linear_expr():
     # pyadjoint/tests/firedrake_adjoint/test_hessian.py::test_nonlinear
     # with modifications where indicated.
 
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
 
     # Get tape instead of creating a new one for consistency with other tests
     tape = get_working_tape()
@@ -329,7 +328,7 @@ def test_interpolate_hessian_nonlinear_expr():
     # pyadjoint/tests/firedrake_adjoint/test_hessian.py::test_nonlinear
     # with modifications where indicated.
 
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
 
     # Get tape instead of creating a new one for consistency with other tests
     tape = get_working_tape()
@@ -387,7 +386,7 @@ def test_interpolate_hessian_nonlinear_expr_multi():
     # pyadjoint/tests/firedrake_adjoint/test_hessian.py::test_nonlinear
     # with modifications where indicated.
 
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
 
     # Get tape instead of creating a new one for consistency with other tests
     tape = get_working_tape()
@@ -451,7 +450,7 @@ def test_ioperator_replay(op, order, power):
     augmented operators +=, -=, *= and /= gives the same
     result as a hand derivation.
     """
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     mesh = UnitSquareMesh(4, 4)
     x, y = SpatialCoordinate(mesh)
     V = FunctionSpace(mesh, "CG", order)
@@ -510,7 +509,7 @@ def supermesh_setup(vector=False):
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_self_supermesh_project():
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     source, target_space = supermesh_setup()
     control = Control(source)
     target = Function(target_space)
@@ -533,7 +532,7 @@ def test_self_supermesh_project():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_supermesh_project_function():
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     source, target_space = supermesh_setup()
     control = Control(source)
     target = Function(target_space)
@@ -556,7 +555,7 @@ def test_supermesh_project_function():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_supermesh_project_to_function_space():
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     source, target_space = supermesh_setup()
     control = Control(source)
     target = project(source, target_space)
@@ -578,7 +577,7 @@ def test_supermesh_project_to_function_space():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_supermesh_project_gradient(vector):
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     source, target_space = supermesh_setup()
     source_space = source.function_space()
     control = Control(source)
@@ -595,7 +594,7 @@ def test_supermesh_project_gradient(vector):
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_supermesh_project_tlm(vector):
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     source, target_space = supermesh_setup()
     control = Control(source)
     target = project(source, target_space)
@@ -616,7 +615,7 @@ def test_supermesh_project_tlm(vector):
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_supermesh_project_hessian(vector):
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     source, target_space = supermesh_setup()
     control = Control(source)
     target = project(source, target_space)
@@ -647,7 +646,7 @@ def test_supermesh_project_hessian(vector):
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_init_constant():
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     mesh = UnitSquareMesh(1, 1)
     c1 = Constant(1.0, domain=mesh)
     c2 = Constant(0.0, domain=mesh)
@@ -659,7 +658,7 @@ def test_init_constant():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_init_constant_diff_mesh():
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     mesh = UnitSquareMesh(1, 1)
     mesh0 = UnitSquareMesh(2, 2)
     c1 = Constant(1.0, domain=mesh)
@@ -672,7 +671,7 @@ def test_init_constant_diff_mesh():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_copy_function():
-    from firedrake_adjoint import ReducedFunctional, Control
+    from firedrake.adjoint import ReducedFunctional, Control
     mesh = UnitSquareMesh(1, 1)
     V = FunctionSpace(mesh, "CG", 1)
     one = Constant(1.0, domain=mesh)
@@ -685,7 +684,7 @@ def test_copy_function():
 
 @pytest.mark.skipcomplex  # Taping for complex-valued 0-forms not yet done
 def test_consecutive_nonlinear_solves():
-    from firedrake_adjoint import ReducedFunctional, Control, taylor_test
+    from firedrake.adjoint import ReducedFunctional, Control, taylor_test
     mesh = UnitSquareMesh(1, 1)
     V = FunctionSpace(mesh, "CG", 1)
     uic = Constant(2.0, domain=mesh)
